refactor: Refactor file structure
