+++ src/routes/questions.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const dataUtils = require('../utils/dataUtils');
+const dataUtils = require('../utils/data.util');
 const source = 'questions.json';
 router.get('/', async function(req, res) {
     const questions = await dataUtils.readFile(source);
